refactor(session): adopt the cleanup renames across the phase-4 surface
The dedup and rename PRs (#736, #737) landed while this branch was in
review: the digest implementation moved to src/lib/reviewDigest, the
broker's resource ceiling became core's reviewResourceCeiling, and the
geometry conformance registry took its dimension name. Point the
phase-4 modules at the new homes instead of re-declaring the old ones.
